+++ src/training/mod.rs
@@ -24,6 +24,7 @@ pub mod shield;
 pub mod tech;
 pub mod throw;
 pub mod ui;
+pub mod ui_hacks;
 
 mod air_dodge_direction;
 mod attack_angle;
